+++ src/BAET/cli/cli.py
@@ -4,11 +4,11 @@ import re
 
 import rich_click as click
 
-from BAET._config.logging import create_logger
-from BAET.cli.help_configuration import BaetConfig
+from BAET._config.logging import configure_logging, create_logger
+from BAET.cli.help_configuration import baet_config
 
-from .command_args import CliOptions, pass_cli_options
-from .commands import extract, probe
+from .commands.extract import extract
+from .commands.probe import probe
 
 logger = create_logger()
 
@@ -16,17 +16,16 @@ file_type_pattern = re.compile(r"^\.?(\w+)$")
 
 
 @click.group()
-@BaetConfig(use_markdown=True)
+@baet_config(use_markdown=True)
 @click.version_option(prog_name="BAET", package_name="BAET", message="%(prog)s v%(version)s")
 @click.option("--logging", "-L", is_flag=True, help="Run the application with logging.")
-@pass_cli_options
-def cli(cli_args: CliOptions, logging: bool) -> None:
+def cli(logging: bool) -> None:
     """**Bulk Audio Extraction Tool (BAET)**
 
     This tool provides a simple way to extract audio from video files.
     - You can use --help on any command to get more information.
     """
-    cli_args.logging = logging
+    configure_logging(enable_logging=logging)
 
 
 cli.add_command(extract)
